refactored mocking code
